Add files via upload
AAL is a household name in the United States, known for its clothing business since 2000. It caters to all groups— kids, women, men, and seniors. It has opened branches in many states, metropolises, and tier-1 and tier-2 cities in the United States. The business is booming, and the company is in expansion mode. The company wants to get an in-depth understanding of its sales performance so that it can make investment decisions.

In this project, I compiled, analyzed, and illustrated data findings using charts, graphs, and other visual representations on the enclosed data to derive meaningful insights and generate comprehensive reports.